+++ src/main/java/DAO/IDAOBook.java
@@ -9,7 +9,7 @@ import java.util.Map;
 
 public interface IDAOBook {
     void addBook(Builder newBook);
-    void editBook(long ISBN, int price);
+    void editBook(long ISBN, float price);
     void deleteBook(long ISBN);
     Map<String,Book> searchBooks(String surname);
     Map<String,Book> getBooks();
